Added Program and Disassembler classes
Added a program, instruction and disassembler class.
