+++ Assets/Scenes/Tomasulo.unity
@@ -181,7 +181,7 @@ MonoBehaviour:
     m_HorizontalOverflow: 0
     m_VerticalOverflow: 0
     m_LineSpacing: 1
-  m_Text: Load Instructions
+  m_Text: Load
 --- !u!222 &189265277
 CanvasRenderer:
   m_ObjectHideFlags: 0
@@ -1184,7 +1184,7 @@ GameObject:
   - component: {fileID: 1878490905}
   - component: {fileID: 1878490908}
   m_Layer: 5
-  m_Name: ButtonLoadInstructions
+  m_Name: ButtonLoadFromDisk
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
